+++ bindings/node/lib/bindings/tokenizer.js
@@ -1,5 +1,6 @@
 const native = require("./native");
 
 module.exports = {
+  AddedToken: native.tokenizer_AddedToken,
   Tokenizer: native.tokenizer_Tokenizer
 };