+++ bindings/node/lib/implementations/tokenizers/bert-wordpiece.tokenizer.ts
@@ -5,9 +5,9 @@ import { Model, WordPiece, WordPieceOptions } from "../../bindings/models";
 import { bertNormalizer } from "../../bindings/normalizers";
 import { bertProcessing } from "../../bindings/post-processors";
 import { bertPreTokenizer } from "../../bindings/pre-tokenizers";
-import { Tokenizer } from "../../bindings/tokenizer";
+import { AddedToken, Tokenizer } from "../../bindings/tokenizer";
 import { wordPieceTrainer } from "../../bindings/trainers";
-import { BaseTokenizer } from "./base.tokenizer";
+import { BaseTokenizer, getTokenContent } from "./base.tokenizer";
 
 export interface BertWordPieceOptions {
   /**
@@ -21,7 +21,7 @@ export interface BertWordPieceOptions {
   /**
    * @default "[CLS]"
    */
-  clsToken?: string;
+  clsToken?: string | AddedToken;
   /**
    * @default true
    */
@@ -33,15 +33,15 @@ export interface BertWordPieceOptions {
   /**
    * @default "[MASK]"
    */
-  maskToken?: string;
+  maskToken?: string | AddedToken;
   /**
    * @default "[PAD]"
    */
-  padToken?: string;
+  padToken?: string | AddedToken;
   /**
    * @default "[SEP]"
    */
-  sepToken?: string;
+  sepToken?: string | AddedToken;
   /**
    * @default true
    */
@@ -49,7 +49,7 @@ export interface BertWordPieceOptions {
   /**
    * @default "[UNK]"
    */
-  unkToken?: string;
+  unkToken?: string | AddedToken;
   vocabFile?: string;
   /**
    * The prefix to attach to subword units that don't represent a beginning of word
@@ -78,7 +78,7 @@ export interface BertWordPieceTrainOptions {
   /**
    * @default ["[PAD]", "[UNK]", "[CLS]", "[SEP]", "[MASK]"]
    */
-  specialTokens?: string[];
+  specialTokens?: (string | AddedToken)[];
   /**
    * @default 30000
    */
@@ -139,7 +139,7 @@ export class BertWordPieceTokenizer extends BaseTokenizer<BertTokenizerConfig> {
     if (opts.vocabFile) {
       const fromFiles = promisify<string, WordPieceOptions, Model>(WordPiece.fromFiles);
       model = await fromFiles(opts.vocabFile, {
-        unkToken: opts.unkToken,
+        unkToken: getTokenContent(opts.unkToken),
         continuingSubwordPrefix: opts.wordpiecesPrefix
       });
     } else {
@@ -155,7 +155,7 @@ export class BertWordPieceTokenizer extends BaseTokenizer<BertTokenizerConfig> {
       opts.padToken,
       opts.maskToken
     ]) {
-      if (tokenizer.tokenToId(token) !== undefined) {
+      if (tokenizer.tokenToId(getTokenContent(token)) !== undefined) {
         tokenizer.addSpecialTokens([token]);
       }
     }
@@ -165,19 +165,19 @@ export class BertWordPieceTokenizer extends BaseTokenizer<BertTokenizerConfig> {
     tokenizer.setPreTokenizer(bertPreTokenizer());
 
     if (opts.vocabFile && opts.addSpecialTokens) {
-      const sepTokenId = tokenizer.tokenToId(opts.sepToken);
+      const sepTokenId = tokenizer.tokenToId(getTokenContent(opts.sepToken));
       if (sepTokenId === undefined) {
         throw new Error("sepToken not found in the vocabulary");
       }
 
-      const clsTokenId = tokenizer.tokenToId(opts.clsToken);
+      const clsTokenId = tokenizer.tokenToId(getTokenContent(opts.clsToken));
       if (clsTokenId === undefined) {
         throw new Error("clsToken not found in the vocabulary");
       }
 
       const processor = bertProcessing(
-        [opts.sepToken, sepTokenId],
-        [opts.clsToken, clsTokenId]
+        [getTokenContent(opts.sepToken), sepTokenId],
+        [getTokenContent(opts.clsToken), clsTokenId]
       );
       tokenizer.setPostProcessor(processor);
     }