+++ bindings/node/lib/implementations/tokenizers/index.ts
@@ -2,3 +2,4 @@ export * from "./bert-wordpiece.tokenizer";
 export * from "./bpe.tokenizer";
 export * from "./byte-level-bpe.tokenizer";
 export * from "./sentence-piece-bpe.tokenizer";
+export { getTokenContent } from "./base.tokenizer";