+++ tokenizers/src/cli.rs
@@ -22,14 +22,8 @@ fn shell(matches: &ArgMatches) -> Result<()> {
     tokenizer.with_decoder(Box::new(ByteLevel::default()));
 
     tokenizer.add_tokens(&[
-        AddedToken {
-            content: String::from("ing"),
-            single_word: false,
-        },
-        AddedToken {
-            content: String::from("[ENT]"),
-            single_word: true,
-        },
+        AddedToken::from(String::from("ing")).single_word(false),
+        AddedToken::from(String::from("[ENT]")).single_word(true),
     ]);
 
     let stdin = io::stdin();