+++ tokenizers/src/tokenizer/mod.rs
@@ -15,7 +15,7 @@ pub use crate::utils::truncation::{truncate_encodings, TruncationParams, Truncat
 use indicatif::{ProgressBar, ProgressStyle};
 use rayon::prelude::*;
 use std::{
-    collections::HashMap,
+    collections::{HashMap, HashSet},
     fs::File,
     io::{BufRead, BufReader},
     path::{Path, PathBuf},
@@ -94,7 +94,10 @@ pub trait Trainer: Sync {
     fn should_show_progress(&self) -> bool;
     /// The actual training method. This will return a new trained Model as well as a list
     /// of `special_tokens` to be added directly to the tokenizer along with the model.
-    fn train(&self, words: HashMap<String, u32>) -> Result<(Box<dyn Model + Sync>, Vec<String>)>;
+    fn train(
+        &self,
+        words: HashMap<String, u32>,
+    ) -> Result<(Box<dyn Model + Sync>, Vec<AddedToken>)>;
     /// Process a bunch of token, counting them as relevant.
     fn process_tokens(&self, words: &mut HashMap<String, u32>, tokens: Vec<String>);
 }
@@ -123,13 +126,70 @@ pub struct AddedToken {
     pub content: String,
     /// Whether this token must be a single word or can break words
     pub single_word: bool,
+    /// Whether this token should strip whitespaces on its left
+    pub lstrip: bool,
+    /// Whether this token should strip whitespaces on its right
+    pub rstrip: bool,
 }
 impl AddedToken {
-    fn from(content: String) -> Self {
+    pub fn from(content: String) -> Self {
         AddedToken {
             content,
             ..Default::default()
         }
+    }
+    pub fn single_word(mut self, single_word: bool) -> Self {
+        self.single_word = single_word;
+        self
+    }
+    pub fn lstrip(mut self, lstrip: bool) -> Self {
+        self.lstrip = lstrip;
+        self
+    }
+    pub fn rstrip(mut self, rstrip: bool) -> Self {
+        self.rstrip = rstrip;
+        self
+    }
+    pub fn get_pattern(&self) -> String {
+        let mut r = if self.single_word {
+            let first_b = self
+                .content
+                .chars()
+                .next()
+                .map(|c| {
+                    if regex_syntax::is_word_character(c) {
+                        r"\b"
+                    } else {
+                        ""
+                    }
+                })
+                .unwrap();
+            let last_b = self
+                .content
+                .chars()
+                .last()
+                .map(|c| {
+                    if regex_syntax::is_word_character(c) {
+                        r"\b"
+                    } else {
+                        ""
+                    }
+                })
+                .unwrap();
+            format!(r"{}{}{}", first_b, regex::escape(&self.content), last_b)
+        } else {
+            regex::escape(&self.content)
+        };
+
+        if self.lstrip && self.rstrip {
+            r = format!(r"(\s)?{}(\s)?", r);
+        } else if self.lstrip {
+            r = format!(r"(\s)?{}", r);
+        } else if self.rstrip {
+            r = format!(r"{}(\s)?", r);
+        }
+
+        r
     }
 }
 impl Default for AddedToken {
@@ -137,6 +197,8 @@ impl Default for AddedToken {
         AddedToken {
             content: String::new(),
             single_word: false,
+            lstrip: false,
+            rstrip: false,
         }
     }
 }
@@ -164,10 +226,21 @@ pub struct Tokenizer {
     decoder: Option<Box<dyn Decoder + Sync>>,
 
     // Added Vocabulary capabilities
-    added_tokens: HashMap<AddedToken, u32>,
-    added_tokens_r: HashMap<u32, AddedToken>,
-    split_re: Option<regex::Regex>,
-    special_tokens: HashMap<String, u32>,
+    /// Contains the mapping from String to ID as the user intended it. This map
+    /// contains both special tokens and classic added tokens.
+    added_tokens_map: HashMap<String, u32>,
+    /// Contains the mapping from ID to AddedToken for all the added tokens, both special
+    /// and classic.
+    added_tokens_map_r: HashMap<u32, AddedToken>,
+    /// Contains only the classic AddedToken, in the specific order the user gave them.
+    added_tokens: Vec<AddedToken>,
+    /// Contains only the special AddedToken, in the specific order the user gave them.
+    special_tokens: Vec<AddedToken>,
+    /// A Set, containing all the special token for easy access while decoding. This let's
+    /// use remove them easily with an O(1) complexity.
+    special_tokens_set: HashSet<String>,
+    /// A RegexSet containing all the patterns used to split on AddedTokens
+    split_re: regex::RegexSet,
 
     // General processing parameters
     trunc: Option<TruncationParams>,
@@ -184,10 +257,12 @@ impl Tokenizer {
             post_processor: None,
             decoder: None,
 
-            added_tokens: HashMap::new(),
-            added_tokens_r: HashMap::new(),
-            split_re: None,
-            special_tokens: HashMap::new(),
+            added_tokens_map: HashMap::new(),
+            added_tokens_map_r: HashMap::new(),
+            added_tokens: vec![],
+            special_tokens: vec![],
+            special_tokens_set: HashSet::new(),
+            split_re: regex::RegexSet::new::<_, &&str>(&[]).unwrap(),
 
             trunc: None,
             padding: None,
@@ -278,7 +353,7 @@ impl Tokenizer {
 
     /// Converts a token in the corresponding id.
     pub fn token_to_id(&self, token: &str) -> Option<u32> {
-        if let Some(id) = self.added_tokens.get(&AddedToken::from(token.to_owned())) {
+        if let Some(id) = self.added_tokens_map.get(token) {
             Some(*id)
         } else {
             self.model.token_to_id(token)
@@ -287,7 +362,7 @@ impl Tokenizer {
 
     /// Converts an id to the corresponding token.
     pub fn id_to_token(&self, id: u32) -> Option<String> {
-        if let Some(token) = self.added_tokens_r.get(&id) {
+        if let Some(token) = self.added_tokens_map_r.get(&id) {
             Some(token.content.clone())
         } else {
             self.model.id_to_token(id)
@@ -301,9 +376,9 @@ impl Tokenizer {
             .into_iter()
             .map(|(sentence, id)| -> Result<NormalizedString> {
                 if id.is_some() {
-                    Ok(NormalizedString::from(&sentence))
+                    Ok(sentence)
                 } else {
-                    let mut normalized = self.do_normalize(&sentence)?;
+                    let mut normalized = self.do_normalize(sentence)?;
                     let _ = self.pre_tokenize(&mut normalized)?;
 
                     Ok(normalized)
@@ -334,18 +409,18 @@ impl Tokenizer {
                                 Encoding::new(
                                     vec![id],
                                     vec![type_id],
-                                    vec![sentence.to_owned()],
+                                    vec![sentence.get().to_owned()],
                                     vec![(0, sentence.len())],
                                     vec![0],
                                     vec![1],
                                     vec![],
                                 ),
-                                NormalizedString::from(&sentence),
+                                sentence,
                             ));
                         }
 
                         // 1. Normalization
-                        let mut normalized = self.do_normalize(&sentence)?;
+                        let mut normalized = self.do_normalize(sentence)?;
 
                         // 2. Pre tokenization
                         let pre_tokenized = self.pre_tokenize(&mut normalized)?;
@@ -390,20 +465,29 @@ impl Tokenizer {
                     return Ok((Encoding::default(), NormalizedString::from("")));
                 }
 
+                // Merge encodings and normalized strings
                 let others = encodings.split_off(1);
-                let mut first: Encoding = encodings.into_iter().next().unwrap();
+                let n_others = normalized.split_off(1);
 
-                for encoding in others {
-                    first.merge_with(encoding, true);
+                let mut final_encoding: Encoding = encodings.into_iter().next().unwrap();
+                let mut final_normalized: NormalizedString = normalized.into_iter().next().unwrap();
+
+                let mut offset = final_normalized.len_original();
+                for (mut encoding, normalized) in others.into_iter().zip(n_others) {
+                    encoding
+                        .get_offsets_mut()
+                        .iter_mut()
+                        .for_each(|(start, end)| {
+                            *start += offset;
+                            *end += offset;
+                        });
+                    offset += normalized.len();
+
+                    final_encoding.merge_with(encoding, false);
+                    final_normalized.merge_with(&normalized);
                 }
 
-                let others = normalized.split_off(1);
-                let mut normalized: NormalizedString = normalized.into_iter().next().unwrap();
-                for n in others {
-                    normalized.merge_with(&n);
-                }
-
-                Ok((first, normalized))
+                Ok((final_encoding, final_normalized))
             };
 
         let (sentence, pair) = match input {
@@ -468,14 +552,14 @@ impl Tokenizer {
         let tokens = ids
             .into_iter()
             .map(|id| {
-                let token = if let Some(token) = self.added_tokens_r.get(&id) {
+                let token = if let Some(token) = self.added_tokens_map_r.get(&id) {
                     Some(token.content.to_owned())
                 } else {
                     self.model.id_to_token(id)
                 };
 
                 token.filter(|token| {
-                    !skip_special_tokens || !self.special_tokens.contains_key(token)
+                    !skip_special_tokens || !self.special_tokens_set.contains(token)
                 })
             })
             .filter(|token| token.is_some())
@@ -530,7 +614,7 @@ impl Tokenizer {
                     match file.read_line(&mut buf)? {
                         0 => break,
                         b => {
-                            let mut normalized = self.do_normalize(&buf)?;
+                            let mut normalized = self.do_normalize(NormalizedString::from(&buf))?;
                             let pre_tokenized = self.pre_tokenize(&mut normalized)?;
                             trainer.process_tokens(
                                 &mut words,
@@ -581,9 +665,7 @@ impl Tokenizer {
     }
 
     /// Normalization logic, go through all normalizers
-    fn do_normalize(&self, sequence: &str) -> Result<NormalizedString> {
-        let mut normalized = NormalizedString::from(sequence);
-
+    fn do_normalize(&self, mut normalized: NormalizedString) -> Result<NormalizedString> {
         if let Some(normalizer) = &self.normalizer {
             normalizer.normalize(&mut normalized)?;
         }
@@ -652,20 +734,14 @@ impl Tokenizer {
 
     /// Register the given tokens as special tokens. This is especially useful for removing
     /// these special tokens while decoding
-    pub fn add_special_tokens<T: AsRef<str>>(&mut self, tokens: &[T]) -> usize {
-        let added_tokens = tokens
-            .iter()
-            .map(|t| AddedToken::from(t.as_ref().to_owned()))
-            .collect::<Vec<_>>();
-
-        let added = self.add_tokens(&added_tokens);
+    pub fn add_special_tokens(&mut self, tokens: &[AddedToken]) -> usize {
         for token in tokens {
-            if let Some(id) = self.token_to_id(token.as_ref()) {
-                self.special_tokens
-                    .entry(token.as_ref().to_owned())
-                    .or_insert(id);
+            if !self.special_tokens_set.contains(&token.content) {
+                self.special_tokens.push(token.to_owned());
+                self.special_tokens_set.insert(token.content.clone());
             }
         }
+        let added = self.add_tokens(&tokens);
 
         self.refresh_added_tokens();
 
@@ -676,21 +752,27 @@ impl Tokenizer {
     pub fn add_tokens(&mut self, tokens: &[AddedToken]) -> usize {
         let mut ignored = 0;
         for token in tokens {
-            if token.content.is_empty() || self.token_to_id(&token.content).is_some() {
+            if token.content.is_empty() {
                 ignored += 1;
                 continue;
             }
 
-            let new_id = (self.model.get_vocab_size() + self.added_tokens.len()) as u32;
-            let id = self
-                .added_tokens
-                .entry(token.clone())
-                .and_modify(|_| ignored += 1)
-                .or_insert(new_id);
+            let id = if let Some(id) = self.token_to_id(&token.content) {
+                id
+            } else {
+                let new_id = (self.model.get_vocab_size() + self.added_tokens.len()) as u32;
+                self.added_tokens_map.insert(token.content.clone(), new_id);
+
+                if !self.special_tokens_set.contains(&token.content) {
+                    self.added_tokens.push(token.clone());
+                }
+
+                new_id
+            };
 
             // Update the current revert operation
-            self.added_tokens_r
-                .entry(*id)
+            self.added_tokens_map_r
+                .entry(id)
                 .and_modify(|t| *t = token.clone())
                 .or_insert_with(|| token.clone());
         }
@@ -702,110 +784,132 @@ impl Tokenizer {
     }
 
     fn refresh_added_tokens(&mut self) {
-        // We rebuild the regex here everytime on purpose, because the added tokens may
-        // have changed
-        let special_tokens = self
-            .special_tokens
-            .keys()
-            .map(|t| AddedToken {
-                content: t.to_owned(),
-                single_word: true,
-            })
-            .collect::<Vec<_>>();
-        let added_tokens = self
-            .added_tokens
-            .keys()
-            .chain(special_tokens.iter())
-            .map(|token| {
-                if token.single_word {
-                    let first_b = token
-                        .content
-                        .chars()
-                        .next()
-                        .map(|c| {
-                            if regex_syntax::is_word_character(c) {
-                                r"\b"
-                            } else {
-                                ""
-                            }
-                        })
-                        .unwrap();
-                    let last_b = token
-                        .content
-                        .chars()
-                        .last()
-                        .map(|c| {
-                            if regex_syntax::is_word_character(c) {
-                                r"\b"
-                            } else {
-                                ""
-                            }
-                        })
-                        .unwrap();
-                    format!(r"{}{}{}", first_b, regex::escape(&token.content), last_b)
-                } else {
-                    regex::escape(&token.content)
-                }
-            })
-            .collect::<Vec<_>>();
-
-        if added_tokens.is_empty() {
-            self.split_re = None;
-        } else {
-            self.split_re =
-                Some(regex::Regex::new(&format!(r"({})", added_tokens.join("|"))).unwrap());
-        }
+        self.split_re = regex::RegexSet::new(
+            self.special_tokens
+                .iter()
+                .chain(self.added_tokens.iter())
+                .map(|token| token.get_pattern()),
+        )
+        .unwrap();
     }
 
-    /// Split the given sentence on multiple parts, finding the added tokens and their id in the process
-    fn split_on_added_tokens(&self, sentence: &str) -> Vec<(String, Option<u32>)> {
-        if let Some(split_re) = &self.split_re {
-            let splits = split_re
-                .find_iter(&sentence)
-                .map(|m| (m.start(), m.end()))
-                .collect::<Vec<_>>();
+    /// Split the given sentence on multiple parts, finding the added tokens and their id in
+    /// the process
+    fn split_on_added_tokens(&self, sentence: &str) -> Vec<(NormalizedString, Option<u32>)> {
+        let mut matches = self
+            .split_re
+            .matches(sentence)
+            .into_iter()
+            .flat_map(|idx| {
+                regex::Regex::new(&self.split_re.patterns()[idx])
+                    .unwrap()
+                    .find_iter(&sentence)
+                    .map(|m| (idx, (m.start(), m.end())))
+                    .collect::<Vec<_>>()
+            })
+            .collect::<Vec<_>>();
 
-            // We also insert the splits that are inbetween the added tokens, to split the entire string
-            let mut start_offset = 0;
-            let mut splits = splits
-                .into_iter()
-                .flat_map(|(start, end)| {
-                    let mut splits = vec![];
-                    if start_offset < start {
-                        splits.push((start_offset, start));
-                    }
-                    splits.push((start, end));
-                    start_offset = end;
+        // We sort all the matches by their start and then by their pattern id
+        matches.sort_by(
+            |(idxa, (sa, _)), (idxb, (sb, _))| {
+                if sa != sb {
+                    sa.cmp(sb)
+                } else {
+                    idxa.cmp(idxb)
+                }
+            },
+        );
 
-                    splits
-                })
-                .collect::<Vec<_>>();
-            if let Some((_, end)) = splits.iter().last().copied() {
-                if end < sentence.len() {
-                    splits.push((end, sentence.len()));
+        // Select the matches (if some are overlapping) we want to keep
+        let mut i = 0;
+        let mut current_offset = 0;
+        let mut splits = Vec::with_capacity(matches.len());
+        while i < matches.len() {
+            let (idx, (start, end)) = matches[i];
+
+            // current match is before the currentt offset, let's skip it
+            if start < current_offset {
+                i += 1;
+                continue;
+            }
+
+            // Find out if we have overlapping neighbors. If so, we keep the one with the lowest
+            // idx, and apply it, then continue. All others will be skipped since `current_offset`
+            // will have been increased
+            if i + 1 < matches.len() {
+                if let Some((idx, (s, e))) = matches[i..]
+                    .iter()
+                    .take_while(|(_, (s, e))| *s < end && start < *e)
+                    .min() // Order on idx first
+                    .copied()
+                {
+                    splits.push((idx, (s, e)));
+                    current_offset = e;
+                    i += 1;
+                    continue;
                 }
             }
 
-            if splits.is_empty() {
-                vec![(sentence.to_owned(), None)]
-            } else {
+            // We didn't find overlapping neighbors, apply ourself
+            splits.push((idx, (start, end)));
+            current_offset = end;
+            i += 1;
+        }
+
+        // We also insert the splits that are inbetween the added tokens, to split the entire string
+        let mut start_offset = 0;
+        let mut splits = splits
+            .into_iter()
+            .flat_map(|(idx, (start, end))| {
+                let mut splits = vec![];
+                if start_offset < start {
+                    splits.push((None, (start_offset, start)));
+                }
+                splits.push((Some(idx), (start, end)));
+                start_offset = end;
+
                 splits
-                    .into_iter()
-                    .map(|(start, end)| unsafe {
-                        let s = sentence.get_unchecked(start..end).to_owned();
-                        let mut id = self.special_tokens.get(&s);
-                        if id.is_none() {
-                            id = self.added_tokens.get(&AddedToken {
-                                content: s.clone(),
-                                ..Default::default()
-                            });
-                        }
-                        (s, id.copied())
-                    })
-                    .collect()
+            })
+            .collect::<Vec<_>>();
+        if let Some((_, (_, end))) = splits.iter().last().copied() {
+            if end < sentence.len() {
+                splits.push((None, (end, sentence.len())));
             }
+        }
+
+        if splits.is_empty() {
+            vec![(NormalizedString::from(sentence), None)]
         } else {
-            vec![(sentence.to_owned(), None)]
+            splits
+                .into_iter()
+                .map(|(idx, (start, end))| unsafe {
+                    let s = sentence.get_unchecked(start..end).to_owned();
+                    let mut normalized = NormalizedString::from(&s);
+
+                    // Find out the associated AddedToken, and its id
+                    let id = if let Some(idx) = idx {
+                        let added = if idx >= self.special_tokens.len() {
+                            &self.added_tokens[idx - self.special_tokens.len()]
+                        } else {
+                            &self.special_tokens[idx]
+                        };
+
+                        if added.lstrip && added.rstrip {
+                            normalized.strip();
+                        } else if added.lstrip {
+                            normalized.lstrip();
+                        } else if added.rstrip {
+                            normalized.rstrip();
+                        }
+
+                        self.token_to_id(&added.content)
+                    } else {
+                        None
+                    };
+
+                    (normalized, id)
+                })
+                .collect()
         }
     }
 }