+++ tokenizers/tests/offsets.rs
@@ -1,25 +1,13 @@
-use tokenizers::models::bpe::BPE;
-use tokenizers::pre_tokenizers::byte_level::ByteLevel;
-use tokenizers::tokenizer::{get_range_of, EncodeInput, Tokenizer};
+mod common;
 
-fn get_byte_level(add_prefix_space: bool, trim_offsets: bool) -> Tokenizer {
-    let mut tokenizer = Tokenizer::new(Box::new(
-        BPE::from_files("data/gpt2-vocab.json", "data/gpt2-merges.txt")
-            .build()
-            .expect("Files not found, run `make test` to download these files"),
-    ));
-    tokenizer.with_pre_tokenizer(Box::new(
-        ByteLevel::default().add_prefix_space(add_prefix_space),
-    ));
-    tokenizer.with_decoder(Box::new(ByteLevel::default()));
-    tokenizer.with_post_processor(Box::new(ByteLevel::default().trim_offsets(trim_offsets)));
+use common::*;
+use tokenizers::tokenizer::{get_range_of, AddedToken, EncodeInput};
 
-    tokenizer
-}
-
-#[inline]
-fn offset_as_range(offset: (usize, usize)) -> std::ops::Range<usize> {
-    offset.0..offset.1
+macro_rules! check_offsets {
+    ($input: expr, $output:expr, $offset:expr, $result:expr) => {
+        let offsets = $output.get_offsets()[$offset];
+        assert_eq!(get_range_of(&$input, offsets.0..offsets.1), Some($result));
+    };
 }
 
 #[test]
@@ -32,29 +20,13 @@ fn byte_level_basic() {
         .encode(EncodeInput::Single(input.clone()), false)
         .unwrap();
 
-    let offsets = output.get_offsets();
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[0])),
-        Some("Hello")
-    );
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[1])),
-        Some(" there")
-    );
-    assert_eq!(get_range_of(&input, offset_as_range(offsets[2])), Some(","));
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[3])),
-        Some(" how")
-    );
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[4])),
-        Some(" are")
-    );
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[5])),
-        Some(" you")
-    );
-    assert_eq!(get_range_of(&input, offset_as_range(offsets[6])), Some("?"));
+    check_offsets!(input, output, 0, "Hello");
+    check_offsets!(input, output, 1, " there");
+    check_offsets!(input, output, 2, ",");
+    check_offsets!(input, output, 3, " how");
+    check_offsets!(input, output, 4, " are");
+    check_offsets!(input, output, 5, " you");
+    check_offsets!(input, output, 6, "?");
 
     // And when trimming offsets:
     let tokenizer = get_byte_level(true, true);
@@ -64,29 +36,13 @@ fn byte_level_basic() {
         .encode(EncodeInput::Single(input.clone()), false)
         .unwrap();
 
-    let offsets = output.get_offsets();
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[0])),
-        Some("Hello")
-    );
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[1])),
-        Some("there")
-    );
-    assert_eq!(get_range_of(&input, offset_as_range(offsets[2])), Some(","));
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[3])),
-        Some("how")
-    );
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[4])),
-        Some("are")
-    );
-    assert_eq!(
-        get_range_of(&input, offset_as_range(offsets[5])),
-        Some("you")
-    );
-    assert_eq!(get_range_of(&input, offset_as_range(offsets[6])), Some("?"));
+    check_offsets!(input, output, 0, "Hello");
+    check_offsets!(input, output, 1, "there");
+    check_offsets!(input, output, 2, ",");
+    check_offsets!(input, output, 3, "how");
+    check_offsets!(input, output, 4, "are");
+    check_offsets!(input, output, 5, "you");
+    check_offsets!(input, output, 6, "?");
 }
 
 #[test]
@@ -98,10 +54,9 @@ fn byte_level_unicode() {
         .encode(EncodeInput::Single(input.clone()), false)
         .unwrap();
 
-    let offsets = output.get_offsets();
-    assert_eq!(get_range_of(&input, offset_as_range(offsets[1])), Some("⭢"));
-    assert_eq!(get_range_of(&input, offset_as_range(offsets[2])), Some("⭢"));
-    assert_eq!(get_range_of(&input, offset_as_range(offsets[3])), Some("⭢"));
+    check_offsets!(input, output, 1, "⭢");
+    check_offsets!(input, output, 2, "⭢");
+    check_offsets!(input, output, 3, "⭢");
 }
 
 #[test]
@@ -149,6 +104,28 @@ fn byte_level_double_sequence() {
             (8, 10),
             (11, 15),
             (15, 16)
+        ]
+    );
+}
+
+#[test]
+fn split_on_added_tokens_bert() {
+    let input = String::from("Yesterday I saw a [MASK] far away");
+
+    let mut tokenizer = get_bert();
+    tokenizer.add_special_tokens(&[AddedToken::from("[MASK]".into())]);
+    let output = tokenizer.encode(EncodeInput::Single(input), false).unwrap();
+
+    assert_eq!(
+        output.get_offsets(),
+        &[
+            (0, 9),
+            (10, 11),
+            (12, 15),
+            (16, 17),
+            (18, 24),
+            (25, 28),
+            (29, 33)
         ]
     );
 }